Major extension module reshuffling
See CHANGES for details, but the brass tacks:

- Anything starting with `T_` and being an extension of a _signature_
has is now suffixed `_exts`;
- Anything else starting with `T_` has split into two versions: one
extending Base, one extending Core.  These now live in `Base_exts` and
`Core_exts` respectively.

This should make it easier to qualify-import or bang-import every
`Travesty` extension of a Jane Street module.  It also means that
`Travesty` itself now only depends on `base`, but it might take a bit of
opam wrangling for this to go through.