Bump to Base 0.16 and OCaml 5.1
Raising OCaml version to be conservative; this is the only version of
OCaml this has been tested against.

Raising Base version has caused every container generic interface to
gain a new phantom type.

This is predominantly a maintenance move; Base 0.16 contains, for
instance, a new signature set for containers that can be mapped and
filtered, and it's more idiomatic to work with those than travesty.